Add logging to assist troubleshooting
