+++ expense-authorization/src/API/EMBC.ExpenseAuthorization.Api/Email/CommunityEmailRecipientsProvider.cs
@@ -1,0 +1,47 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using FlatFiles;
+using FlatFiles.TypeMapping;
+using Serilog;
+
+namespace EMBC.ExpenseAuthorization.Api.Email
+{
+    public class CommunityEmailRecipientsProvider
+    {
+        private readonly ILogger _logger = Log.ForContext<CommunityEmailRecipientsProvider>();
+        
+        public IList<CommunityEmailRecipient> GetCommunityEmailRecipients(string filename)
+        {
+            try
+            {
+                var mapper = SeparatedValueTypeMapper.Define<CommunityEmailRecipient>();
+                mapper.Property(c => c.Community).ColumnName("community");
+                mapper.Property(c => c.To).ColumnName("to");
+                mapper.Property(c => c.Cc).ColumnName("cc");
+                mapper.Property(c => c.Bcc).ColumnName("bcc");
+
+                if (File.Exists(filename))
+                {
+                    using (var reader = new StreamReader(File.OpenRead(filename)))
+                    {
+                        var options = new SeparatedValueOptions { IsFirstRecordSchema = true };
+                        List<CommunityEmailRecipient> recipients = mapper.Read(reader, options).ToList();
+                        _logger.Debug("Found {Count} community email recipient from {Filename}", recipients.Count, filename);
+
+                        return recipients;
+                    }
+                }
+
+                _logger.Warning("Mapping {Filename} not found, returning empty result", filename);
+                return Array.Empty<CommunityEmailRecipient>();
+            }
+            catch (Exception e)
+            {
+                _logger.Error(e, "Failed to read file {Filename}, returning empty result", filename);
+                return Array.Empty<CommunityEmailRecipient>();
+            }
+        }
+    }
+}
